Restructured profile compoenents
