+++ diabetes-dashboard/src/components/EmotionSettings.vue
