+++ diabetes-dashboard/src/components/FoodSettings.vue
