+++ diabetes-dashboard/src/components/UploadData.vue
